changed Navbar to class component and hold the state of login and pass props to NavHome which is changed to f-component
