+++ src/components/navigation/Navbar.js
@@ -1,4 +1,4 @@
-import React, { Fragment }  from 'react'
+import React, { Component, Fragment }  from 'react'
 import styled from 'styled-components'
 import { NavLink, BrowserRouter, Route } from 'react-router-dom'
 import AppBar from '@material-ui/core/AppBar'
@@ -7,6 +7,7 @@ import Toolbar from '@material-ui/core/Toolbar'
 import NavHome from './NavHome'
 import LoginForm from '../userAuth/LoginForm'
 import SignupForm from '../userAuth/SignupForm'
+import firebaseApp from '../../firebase'
 
 const ImgWrapper = styled.span`
   padding: 5px 5px 2px 5px;
@@ -26,26 +27,47 @@ const BarStyle = {
 
 const peruEffects = { textShadow: '0px 0px 15px peru, 0px 0px 15px peru' }
 
-const Navbar = () => {
-  return (
-    <BrowserRouter>
-      <Fragment>
-        <AppBar position="fixed" style={BarStyle}>
-          <Toolbar>
-            <ImgWrapper>
-              <img src={require('../../img/ipark_logo.png')} alt='missing image' />
-            </ImgWrapper>
-            <Button component={NavLink} exact to='/' style={BtnStyle} activeStyle={peruEffects}>Parking Status</Button>
-            <Button component={NavLink} to='/login' style={BtnStyle} activeStyle={peruEffects}>Login</Button>
-            <Button component={NavLink} to='/signup' style={BtnStyle} activeStyle={peruEffects}>Signup</Button>
-          </Toolbar>
-        </AppBar>
-        <Route exact path='/' component={NavHome} />
-        <Route path='/login' component={LoginForm} />
-        <Route path='/signup' component={SignupForm} />
-      </Fragment>
-    </BrowserRouter>
-  )
-}
+let unsubscribeAuth
+const auth = firebaseApp.auth()
 
-export default Navbar
+export default class Navbar extends Component {
+  state = {
+    login: null
+  }
+
+  componentDidMount() {
+    unsubscribeAuth = auth.onAuthStateChanged(user => {
+      console.log(user)
+      const isLogin = user ? true : false
+      this.setState({ login: isLogin })
+      console.log('login value is', this.state.login)
+    })
+  }
+
+  componentWillUnmount() {
+    unsubscribeAuth()
+    console.log('Navbar unmounted')
+  }
+
+  render() {
+    return (
+      <BrowserRouter>
+        <Fragment>
+          <AppBar position="fixed" style={BarStyle}>
+            <Toolbar>
+              <ImgWrapper>
+                <img src={require('../../img/ipark_logo.png')} alt='missing image' />
+              </ImgWrapper>
+              <Button component={NavLink} exact to='/' style={BtnStyle} activeStyle={peruEffects}>Parking Status</Button>
+              <Button component={NavLink} to='/login' style={BtnStyle} activeStyle={peruEffects}>Login</Button>
+              <Button component={NavLink} to='/signup' style={BtnStyle} activeStyle={peruEffects}>Signup</Button>
+            </Toolbar>
+          </AppBar>
+          <Route exact path='/' render={() => <NavHome login={this.state.login} />} />
+          <Route path='/login' component={LoginForm} />
+          <Route path='/signup' component={SignupForm} />
+        </Fragment>
+      </BrowserRouter>
+    )
+  }
+}
